refactor: ytmdClient and touchPortalClient refactor
- refactored `ytmdClient.ts`
  - changed construtor to have host and port passed in as parameters.
  - rewrote how the `syncNewAuthToken` handles getting a new token.

- refactored `touchPortalClient.ts`
  - extracted the constants into `tpConstants.ts`
  - changed return type of the client to `InstanceType<typeof
    TouchPortalApi.Client>`
