+++ src/services/touchPortalClient.ts
@@ -1,35 +1,7 @@
 import * as TouchPortalApi from 'touchportal-api';
+import { PLUGIN_ID } from './utils/tpConstants';
 
-export const TP_ACTIONS = {
-  ytmdPlayPause: "ytmd.action.play/pause",
-  ytmdNextTrack: "ytmd.action.nextTrack",
-  ytmdPrevTrack: "ytmd.action.prevTrack",
-  ytmdMuteUnmute: "ytmd.action.mute/unmute",
-  ytmdVolumeUp: "ytmd.action.volumeUp",
-  ytmdVolumeDown: "ytmd.action.volumeDown",
-  ytmdRepeatMode: "ytmd.action.repeatMode",
-  ytmdShuffleMode: "ytmd.action.shuffleMode",
-  ytmdLikeDislike: "ytmd.action.like/dislike",
-};
-
-export const TP_EVENTS = {
-  ytmdEventPlayback: "ytmd.event.playback",
-  ytmdEventAudioPlayback: "ytmd.event.audioPlayback",
-  ytmdEventRepeatMode: "ytmd.event.repeatMode",
-  ytmdEventShuffleMode: "ytmd.event.shuffleMode",
-  ytmdEventLikeDislike: "ytmd.event.likeDislike",
-};
-
-export const TP_STATES = {
-  ytmdPlaybackState: "ytmd.states.playbackState",
-  ytmdMutedState: "ytmd.states.mutedState",
-  ytmdRepeatMode: "ytmd.states.repeatMode",
-  ytmdShuffleMode: "ytmd.states.shuffleMode",
-  ytmdLikeDislike: "ytmd.states.likeDislike",
-};
-
-const PLUGIN_ID = 'ytmd_plugin_v2';
-export async function touchPortalClient(): Promise<any> {
+export async function touchPortalClient(): Promise<InstanceType<typeof TouchPortalApi.Client>> {
   console.log('Initializing TouchPortal client...');
   
   const touchPortalClient = new TouchPortalApi.Client();
