+++ src/services/ytmdClient.ts
@@ -3,12 +3,14 @@ import path from 'path';
 import { CompanionConnector, RestClient, Settings, SocketClient } from 'ytmdesktop-ts-companion';
 
 const DEFAULT_DELAY = 2000;
+const MAX_ATTEMPTS = 15;
 export class YTMDClient {
   private companionConnector: CompanionConnector;
   public restClient: RestClient;
   public socketClient: SocketClient;
   private tokenPath: string;
-  constructor() {
+  
+  constructor(host: string = "127.0.0.1", port: number = 9863) {
     const exeDir = path.dirname(process.execPath);
     const pluginRootDir = path.join(exeDir, '.');
     
@@ -16,8 +18,8 @@ export class YTMDClient {
     
     const version = this.getVersion();
     const settings: Settings = {
-      host: "127.0.0.1",
-      port: 9863,
+      host: host,
+      port: port,
       appId: "touchportal-ytmdesktop-plugin",
       appName: "TouchPortal YTMDesktop Plugin",
       appVersion: version,
@@ -62,35 +64,50 @@ export class YTMDClient {
 
       let token = null;
       let attempts = 0;
-      const maxAttempts = 15;
 
-      while (!token && attempts < maxAttempts) {
+      /**
+       * Wait for the user to click the token modal.
+       * YTMD closes the modal after 30 seconds.
+       */
+      while (!token && attempts < MAX_ATTEMPTS) {
         try {
           await this.delay(DEFAULT_DELAY);
           //get token
           const tokenResponse = await this.restClient.getAuthToken(
             response.code,
           );
+          
           token = tokenResponse.token;
+          this.companionConnector.setAuthToken(token);
+          this.saveToken(token);
+          return;
         } catch (error) {
           attempts++;
         }
       }
 
-      if (!token) {
-        throw new Error(
-          "Authentication Timeout. User did not click 'Allow' in time.",
-        );
-      }
-
-      this.companionConnector.setAuthToken(token);
-      fs.writeFileSync(this.tokenPath, token, "utf-8");
+      throw new Error(
+        "Authentication Timeout. User did not click 'Allow' in time.",
+      );
+      
+      // fs.writeFileSync(this.tokenPath, token, "utf-8");
     } catch (error) {
       console.error("Error during auth token synchronization:", error);
       throw error;
     }
   }
 
+  private saveToken(token: string): void {
+    try {
+      if (token) {
+      fs.writeFileSync(token, token, "utf-8");
+    }
+    } catch (error) {
+      console.log("");
+    }
+    
+  }
+  
   private getToken(): string | null {
     try {
       const token = fs.readFileSync(this.tokenPath, "utf-8").trim();
